Fix some bugs of create database.

+++ CNNCStorageDB/Models/Project.cs
@@ -15,10 +15,10 @@ namespace CNNCStorageDB.Models
         }
         public int Id { get; set; }
         public string Name { get; set; }
-        public decimal Budget { get; set; }
+        public int Budget { get; set; }
         public DateTime StartTime { get; set; }
         public DateTime EndTime { get; set; }
-        public decimal FinalPrice { get; set; }
+        public double FinalPrice { get; set; }
 
         public ICollection<Client> Clients { get; set; }
         public ICollection<Worker> Workers { get; set; }
